Move helpers under testing/pytest
