+++ localstack/testing/pytest/cloudtrail_tracking/.gitignore
