+++ localstack/testing/pytest/cloudtrail_tracking/cdk.json
