+++ localstack/testing/pytest/cloudtrail_tracking/cloudtrail_tracking/cloudtrail_tracking_stack.py
