+++ localstack/testing/pytest/cloudtrail_tracking/cloudtrail_tracking/handler/index.py
