+++ localstack/testing/pytest/cloudtrail_tracking/requirements.txt
