+++ localstack/testing/pytest/cloudtrail_tracking/tests/test_cloudtrail_tracking_handler.py
